Liquidation UT

+++ contracts/core/tokens/IGearingToken.sol
@@ -56,7 +56,7 @@ interface IGearingToken is IERC721 {
     /// @notice Error for the ltv increase after liquidation
     /// @param ltvBefore Loan to value before liquidation
     /// @param ltvAfter Loan to value after liquidation
-    error LtvIncreasedAfterLiquidation(uint256 ltvBefore, uint256 ltvAfter);
+    error LtvIncreasedAfterLiquidation(uint128 ltvBefore, uint128 ltvAfter);
     /// @notice Error for unauthorized operation
     /// @param id The id of Gearing Token
     error CallerIsNotTheOwner(uint256 id);
